+++ core/server/services/members/service.js
@@ -4,7 +4,7 @@ const MembersSSR = require('@tryghost/members-ssr');
 const db = require('../../data/db');
 const MembersConfigProvider = require('./config');
 const MembersCSVImporter = require('@tryghost/members-importer');
-const MembersStats = require('./stats');
+const MembersStats = require('./stats/members-stats');
 const createMembersApiInstance = require('./api');
 const createMembersSettingsInstance = require('./settings');
 const logging = require('@tryghost/logging');
